修改 baseMapper selectCount int -> long Mybatis Plus 在3.4 版本之后将 selectCount 从Integer 改为Long

+++ yudao-module-infra/yudao-module-infra-impl/src/main/java/cn/iocoder/yudao/module/infra/dal/mysql/file/FileMapper.java
@@ -24,7 +24,7 @@ public interface FileMapper extends BaseMapperX<FileDO> {
                 .orderByDesc("create_time"));
     }
 
-    default Integer selectCountById(String id) {
+    default Long selectCountById(String id) {
         return selectCount(FileDO::getId, id);
     }
